+++ modules/AWSPowerShell/Cmdlets/AutoScaling/Basic/New-ASAutoScalingGroup-Cmdlet.cs
@@ -202,10 +202,10 @@ namespace Amazon.PowerShell.Cmdlets.AS
         #region Parameter HealthCheckType
         /// <summary>
         /// <para>
-        /// <para>A comma-separated list of one or more health check types.</para><para>The valid values are <code>EC2</code>, <code>ELB</code>, and <code>VPC_LATTICE</code>.
+        /// <para>A comma-separated value string of one or more health check types.</para><para>The valid values are <code>EC2</code>, <code>ELB</code>, and <code>VPC_LATTICE</code>.
         /// <code>EC2</code> is the default health check and cannot be disabled. For more information,
         /// see <a href="https://docs.aws.amazon.com/autoscaling/ec2/userguide/healthcheck.html">Health
-        /// checks for Auto Scaling instances</a> in the <i>Amazon EC2 Auto Scaling User Guide</i>.</para>
+        /// checks for Auto Scaling instances</a> in the <i>Amazon EC2 Auto Scaling User Guide</i>.</para><para>Only specify <code>EC2</code> if you must clear a value that was previously set.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]